fix(api,web): three "she said she did it" bugs, per-day costs, and Bali's OSM pool (v10.71.0)
Validated every claim in docs/video-script-4min.md against live production
rather than reading it, ahead of recording the demo. Findings, reproducible
test cases with verbatim outputs, and a recording runbook are in the new
docs/demo-script-validation.md. Three of the failures shared one shape: Anya
narrating an action she has no way to perform.

1. A claimed pin that never existed. "I really want to see Tanah Lot on this
   trip" returned named_interest: null and pinned_pois: [] while replying "Got
   it! I've added Tanah Lot to your pinned points of interest." Pinning is
   driven by named interests, so a bare place name ran no pipeline at all.
   Fixed at both layers: the refine prompt routes a specific named place into
   named_interest so it reaches verification, and _strip_false_pin_claim /
   _HALLUCINATED_PIN_RE replace any past-tense pin claim in a turn that pinned
   nothing -- the same structural guard, for the same reason, as
   wizard_chat_chain's _HALLUCINATED_GENERATION_RE. Anchored on past/perfect
   forms only, so "I'll add", "Shall I pin" and the genuinely-performed "I've
   updated your budget" all pass untouched.

2. The honesty message was unreachable. _apply_interest_pinning returned early
   when interest expansion produced no candidates, skipping the message written
   for exactly that case. Live, "Harry Potter" in Goa yields zero candidates,
   so the user got "I'll look for real, verified Harry Potter places" and no
   follow-up -- an unkept promise, which is worse than the invention the
   pipeline exists to prevent. Expansion finding nothing and everything failing
   verification are the same event to a traveller and now produce the same
   reply, asserted byte-identical by test.

3. "make day 3 cheaper" promised an edit that never happened -- action_type
   "none", no patch, no regeneration, no diff. There was nothing to move: cost
   existed only as one whole-trip ExpenseBreakdown, with no per-item or per-day
   figure anywhere. Now a real feature in three layers:
   - ItineraryItem.estimated_cost_inr (group cost of one item; excludes
     flights/accommodation, which are trip-level) with a derived
     ItineraryDay.estimated_cost_inr property, so a day total cannot drift from
     the items it summarises.
   - TripConfig.day_cost_preferences, structured as day + Literal["cheaper",
     "pricier"] rather than free text, deduped last-write-wins so iterative
     refinement cannot leave one day both.
   - _day_cost_guidance_block in the generation prompt, plus client-side
     regeneration-and-diff on the same path pins already use.
   The day number is parsed deterministically from the user's own text, not
   trusted from the model: it is a structural index into the itinerary, and a
   model that miscounts it silently re-costs the wrong day.

Cost sanity, and the one retry it earns. Live runs exposed two failure modes
invisible to every existing check, because the itinerary is otherwise perfect
-- right day count, real places, pins honoured, internally consistent totals.
Wrong currency: one run in five returned Rs 124,525,000 for a 6-day trip on a
Rs 2.4L budget, Gemini costing in Indonesian Rupiah despite the schema saying
INR, and it corrupted expense_breakdown too, so it is not specific to the new
field. Wrong direction: a day asked to be made cheaper coming back at or above
the other days, at a scale that looks entirely normal. _cost_sanity_problem
catches both and triggers exactly one regeneration carrying a correction that
names the defect and re-anchors every figure to INR. Three deliberate
properties: the check runs BEFORE the cache write (caching a wrong-currency
itinerary would serve the defect to every later fallback for that trip shape);
a second bad answer does not replace the first; and a surviving problem becomes
a user-visible warnings entry rather than being swallowed. The scale anchor is
per-person-per-day and deliberately budget-free -- a stated budget cannot be
the yardstick when the feasibility gate exists precisely for trips that cost
far more than the user typed.

Bali's OSM pool was in the wrong half of the island. 25 stored POIs with a
centroid 48 km from Denpasar, in Buleleng on the north coast, named "Indian
cousin", "Gado gado", "FREE waterfall entrance!". Root cause: a bare "Bali"
query returns the island centroid, and geocode_city corrects that to the hub
town via _hub_town_in_bbox -- which is itself an Overpass call, so when
Overpass throttles it silently falls back to the raw centroid. Confirmed live:
it returned Denpasar cleanly, then 429, then 504 within one session, and the
stored data's centroid sat 3 km from the fallback point. Ingestion quality
depended on whether an unrelated API happened to be up, and no downstream guard
notices -- count, category share and the prominence check all look healthy.
Pinned with GEOCODE_QUERY_OVERRIDES["bali"].

Separately, the 5 km default radius encodes a city-shaped assumption that is
false for a region: from Denpasar, Tanah Lot is 15.3 km, Ubud 19 km and Uluwatu
22.7 km, so 5 km returned zero marquee landmarks and filled 60 slots with
Denpasar churches, cinemas and gyms. New _OSM_RADIUS_OVERRIDES_M is read inside
ingest_osm_pois, because core/scheduler.py::_refresh_osm_pois calls it bare and
a call-site radius would be silently reverted by the next scheduled refresh;
the thin-destination retry gained a max() guard so it widens from the override
instead of narrowing 30 km back to 15 km. Result: 60 POIs, centroid 0.7 km from
Denpasar, top-category share 0.25, and Tanah Lot, Uluwatu, Sacred Monkey Forest
and Ubud Palace now pin via OSM with real coordinates instead of wiki-only with
lat/lon 0.0 -- so they render on the map.

A lesson recorded because it cost real time: the 30 km radius was dry-run,
showed no Tanah Lot, and was reported as "widening will not help". That run
predated the geocode pin, so it measured a circle around the island centroid,
from which Tanah Lot is 44 km away. The null result was an artifact of the very
bug under investigation. A measurement taken while a known bug is active is not
evidence.

Also: core/domestic_transport_pricing.py had CI mypy red at HEAD (arrived with
v10.70) -- round(x, -1) returns a float, so max(int, float) widened three
-> int returns.

docs/video-script-4min.md corrected for what the app actually does: the gate
needs four wizard turns, not one; the app dictates the budget chip's figure, so
"raise to Rs 1,60,000" was both the wrong label and a figure that would fail the
gate again; seven cost rows render for Bali because zero-value rows are hidden;
one human-handoff trigger exists, not two; and the "TTS is off" setup line was
stale -- Anya's voice has been live since v10.68/69.

Suite 1224 passed / 6 skipped (was 1154), ruff and mypy clean (227 files),
frontend tsc clean apart from a pre-existing useVoice.test.tsx error that
arrived with v10.68.

+++ apps/api/core/domestic_transport_pricing.py
@@ -75,7 +75,11 @@ def _rail_fare_inr(distance_km: float, class_tier: str) -> int:
             _TELESCOPIC_TAPER_KM * rate
             + (distance_km - _TELESCOPIC_TAPER_KM) * rate * _TELESCOPIC_TAPER_DISCOUNT
         )
-    return max(_RAIL_MINIMUM_FARE_INR, round(fare, -1))  # round to nearest 10
+    # `round(x, -1)` on a float returns a float, so `max(int, float)` widens the
+    # whole expression to float and breaks the `-> int` contract these fares are
+    # consumed under (they land in int-typed INR fields). The value is already a
+    # whole multiple of 10 by then, so `int()` is exact, never a truncation.
+    return int(max(_RAIL_MINIMUM_FARE_INR, round(fare, -1)))  # round to nearest 10
 
 
 # ---------------------------------------------------------------------------
@@ -103,7 +107,7 @@ def _bus_fare_inr(distance_km: float, class_tier: str) -> int | None:
     if distance_km > _BUS_MAX_KM:
         return None
     rate = BUS_RATE_PER_KM_INR[class_tier]
-    return max(_BUS_MINIMUM_FARE_INR, round(distance_km * rate, -1))
+    return int(max(_BUS_MINIMUM_FARE_INR, round(distance_km * rate, -1)))
 
 
 # ---------------------------------------------------------------------------
@@ -119,7 +123,7 @@ _CAB_MINIMUM_FARE_INR = 800
 def _cab_fare_inr(distance_km: float) -> int | None:
     if distance_km > CAB_MAX_KM:
         return None
-    return max(_CAB_MINIMUM_FARE_INR, round(distance_km * _CAB_RATE_PER_KM_INR, -1))
+    return int(max(_CAB_MINIMUM_FARE_INR, round(distance_km * _CAB_RATE_PER_KM_INR, -1)))
 
 
 def estimate_domestic_alternative(distance_km: float, class_tier: str) -> dict[str, int | None]:
